requirements.txt
cachetools==1.1.5
dateutils==0.6.6
HTMLParser==0.0.2
python-dateutil==2.4.2
pytz==2015.7
requests==2.8.1
six==1.10.0
spotipy==2.3.7
wheel==0.24.0
xmltodict==0.9.2
